+++ hardware/src/lane/lane_sequencer.sv
@@ -138,6 +138,7 @@ module lane_sequencer import ara_pkg::*; import rvv_pkg::*; import cf_math_pkg::
           VFU_Alu      : pe_req_ready_o = !(operand_request_valid_o[AluA] || operand_request_valid_o[AluB] || operand_request_valid_o[MaskM]);
           VFU_MFpu     : pe_req_ready_o = !(operand_request_valid_o[MulFPUA] || operand_request_valid_o[MulFPUB] || operand_request_valid_o[MulFPUC] || operand_request_valid_o[MaskM]);
           VFU_LoadUnit : pe_req_ready_o = !(operand_request_valid_o[MaskM]);
+          VFU_SlideUnit: pe_req_ready_o = !(operand_request_valid_o[SlideAddrGenA]);
           VFU_StoreUnit: pe_req_ready_o = !(operand_request_valid_o[StA] || operand_request_valid_o[MaskM]);
           VFU_MaskUnit : pe_req_ready_o = !(operand_request_valid_o[AluA] || operand_request_valid_o[AluB] || operand_request_valid_o[MaskB] || operand_request_valid_o[MaskM]);
           default:;
@@ -259,8 +260,8 @@ module lane_sequencer import ara_pkg::*; import rvv_pkg::*; import cf_math_pkg::
 
             operand_request_i[MulFPUC] = '{
               id     : pe_req_i.id,
-              vs     : pe_req_i.swap_vs2_vd_op ? pe_req_i.vs2     : pe_req_i.vd,
-              eew    : pe_req_i.swap_vs2_vd_op ? pe_req_i.eew_vs2 : pe_req_i.eew_vd_op,
+              vs     : pe_req_i.swap_vs2_vd_op ? pe_req_i.vs2            : pe_req_i.vd,
+              eew    : pe_req_i.swap_vs2_vd_op ? pe_req_i.eew_vs2        : pe_req_i.eew_vd_op,
               conv   : pe_req_i.swap_vs2_vd_op ? pe_req_i.conversion_vs2 : OpQueueConversionNone,
               vl     : vfu_operation_d.vl,
               vstart : vfu_operation_d.vstart,
@@ -339,6 +340,85 @@ module lane_sequencer import ara_pkg::*; import rvv_pkg::*; import cf_math_pkg::
             if ((operand_request_i[MaskM].vl << (int'(EW64) - int'(pe_req_i.vtype.vsew))) * NrLanes * 8 != pe_req_i.vl)
               operand_request_i[MaskM].vl += 1;
             operand_request_push[MaskM] = !pe_req_i.vm;
+          end
+          VFU_SlideUnit: begin
+            operand_request_i[SlideAddrGenA] = '{
+              id     : pe_req_i.id,
+              vs     : pe_req_i.vs2,
+              eew    : pe_req_i.eew_vs2,
+              conv   : pe_req_i.conversion_vs2,
+              vtype  : pe_req_i.vtype,
+              vstart : vfu_operation_d.vstart,
+              hazard : pe_req_i.hazard_vs2 | pe_req_i.hazard_vd,
+              default: '0
+            };
+            operand_request_push[SlideAddrGenA] = pe_req_i.use_vs2;
+
+            case (pe_req_i.op)
+              VSLIDEUP: begin
+                // We need to trim full words from the end of the vector that are not used
+                // as operands by the slide unit.
+                automatic vlen_t vslideup_adj = pe_req_i.stride >> ($clog2(NrLanes) + int'(EW64) - int'(pe_req_i.eew_vs2));
+
+                // Since this request goes outside of the lane, we might need to request an
+                // extra operand regardless of whether it is valid in this lane or not.
+                operand_request_i[SlideAddrGenA].vl = pe_req_i.vl / NrLanes - vslideup_adj;
+
+                if ((operand_request_i[SlideAddrGenA].vl + vslideup_adj) * NrLanes != pe_req_i.vl)
+                  operand_request_i[SlideAddrGenA].vl += 1;
+              end
+              VSLIDEDOWN: begin
+                // We need to trim full words from the start of the vector that are not used
+                // as operands by the slide unit.
+                automatic vlen_t vslidedown_adj         = pe_req_i.stride >> ($clog2(NrLanes) + int'(EW64) - int'(pe_req_i.eew_vs2));
+                operand_request_i[SlideAddrGenA].vstart = vslidedown_adj;
+
+                // Since this request goes outside of the lane, we might need to request an
+                // extra operand regardless of whether it is valid in this lane or not.
+                operand_request_i[SlideAddrGenA].vl = pe_req_i.vl / NrLanes;
+                if (operand_request_i[SlideAddrGenA].vl * NrLanes != pe_req_i.vl)
+                  operand_request_i[SlideAddrGenA].vl += 1;
+
+                // If the vslidedown stride is not a full VRF word, we will need to request an extra word
+                if (!pe_req_i.use_scalar_op)
+                  if (pe_req_i.stride - (vslidedown_adj << ($clog2(NrLanes) + int'(EW64) - int'(pe_req_i.eew_vs2))) != 0)
+                    operand_request_i[SlideAddrGenA].vl += 1;
+              end
+            endcase
+
+            // This vector instruction uses masks
+            operand_request_i[MaskM] = '{
+              id     : pe_req_i.id,
+              vs     : VMASK,
+              eew    : pe_req_i.vtype.vsew,
+              vtype  : pe_req_i.vtype,
+              vstart : vfu_operation_d.vstart,
+              hazard : pe_req_i.hazard_vm | pe_req_i.hazard_vd,
+              default: '0
+            };
+            operand_request_push[MaskM] = !pe_req_i.vm;
+
+            case (pe_req_i.op)
+              VSLIDEUP: begin
+                // We need to trim full words from the end of the vector that are not used
+                // as operands by the slide unit.
+                automatic vlen_t vslideup_adj = (pe_req_i.stride / NrLanes / 8) >> (int'(EW64) - int'(pe_req_i.vtype.vsew));
+
+                // Since this request goes outside of the lane, we might need to request an
+                // extra operand regardless of whether it is valid in this lane or not.
+                operand_request_i[MaskM].vl = ((pe_req_i.vl / NrLanes / 8) >> (int'(EW64) - int'(pe_req_i.vtype.vsew))) - vslideup_adj;
+
+                if (((operand_request_i[MaskM].vl + vslideup_adj) << (int'(EW64) - int'(pe_req_i.vtype.vsew))) * NrLanes * 8 != pe_req_i.vl)
+                  operand_request_i[MaskM].vl += 1;
+              end
+              VSLIDEDOWN: begin
+                // Since this request goes outside of the lane, we might need to request an
+                // extra operand regardless of whether it is valid in this lane or not.
+                operand_request_i[MaskM].vl = ((pe_req_i.vl / NrLanes / 8) >> (int'(EW64) - int'(pe_req_i.vtype.vsew)));
+                if ((operand_request_i[MaskM].vl << (int'(EW64) - int'(pe_req_i.vtype.vsew))) * NrLanes * 8 != pe_req_i.vl)
+                  operand_request_i[MaskM].vl += 1;
+              end
+            endcase
           end
           VFU_MaskUnit: begin
             operand_request_i[AluA] = '{